requirements.txt
beautifulsoup4==4.12.3
dominate==2.9.1
hydra-core==1.3.2
matplotlib==3.8.3
numpy==1.26.4
omegaconf==2.3.0
packaging==23.2
Pillow==10.2.0
Requests==2.31.0
scikit-image
torch==2.2.0
torchvision==0.17.0
tqdm==4.65.0
visdom==0.2.4
